requirements.txt
flask
flask-cors
pymongo
bcrypt
pyjwt
python-dotenv